Add initial version of the Welcome and Complete pages
The two pages are implemented as subclasses of a new Page class,
which provides various convenience methods for creating sections
etc. and will in the future also provide the general interface
to propagate results from one page to another.

This commit also introduces a util.lookup.Lookup class to load
files residing the installer directories.

Also add the Baserock logo to the installer, accompanied by a
copyright notice in README.md.