Add more pages: download releases and select a release
Together, these add support for taking the release repository
entered on the welcome page, cloning it into a local cache or
updating it if is already cached, parsing the releases in the
repository and providing the user with a way to select a release
for the installation.

The installation itself is not yet supported.