Add page for downloading release files, add FileDownload class
The FileDownload class comes from InstallIt, my old source-based
Xfce installer that worked similar to the Baserock installer in that
it downloads releases from an online repository.

The DownloadReleasePage class takes the information filled in via
the previous pages and starts downloading all files associated with
the release. At the moment, this is the main image plus any staging
fillers.

Note: Downloads cannot be cancelled at the moment.