Improve threading/progress in download pages, make them cancellable
The release download page now cancels during the prepare stage,
which means that when you go back from it, it'll keep on downloading
but once you forward to it again, it'll restart the download cleanly.

This could obviously be improved to just continue to download unless
the release to be installed has changed since last time the page was
open.